orchestrator/requirements.txt
python-dotenv~=1.0.0
requests~=2.31.0
transformers~=4.35.2
#petals
langchain~=0.0.338
google-generativeai
langchain-google-genai # gemini
langchainhub
langchain_experimental
pillow

# websearch:
selenium~=4.15.2
playwright~=1.45.0
wolframalpha~=5.0.0

face_recognition
opencv-python
git+https://github.com/ageitgey/face_recognition_models

fastapi
aiofiles
streamlit
pipimport3
pyserial

